Module google_datalabeling1_beta1::api

Structs

DataLabeling
Central instance to access all DataLabeling related resource activities

GoogleCloudDatalabelingV1beta1AnnotatedDataset
AnnotatedDataset is a set holding annotations for data in a Dataset. Each labeling task will generate an AnnotatedDataset under the Dataset that the task is requested for.

GoogleCloudDatalabelingV1beta1AnnotatedDatasetMetadata
Metadata on AnnotatedDataset.

GoogleCloudDatalabelingV1beta1Annotation
Annotation for Example. Each example may have one or more annotations. For example in image classification problem, each image might have one or more labels. We call labels binded with this image an Annotation.

GoogleCloudDatalabelingV1beta1AnnotationMetadata
Additional information associated with the annotation.

GoogleCloudDatalabelingV1beta1AnnotationSpec
Container of information related to one possible annotation that can be used in a labeling task. For example, an image classification task where images are labeled as dog or cat must reference an AnnotationSpec for dog and an AnnotationSpec for cat.

GoogleCloudDatalabelingV1beta1AnnotationSpecSet
An AnnotationSpecSet is a collection of label definitions. For example, in image classification tasks, you define a set of possible labels for images as an AnnotationSpecSet. An AnnotationSpecSet is immutable upon creation.

GoogleCloudDatalabelingV1beta1AnnotationSpecSetConfig
Annotation spec set with the setting of allowing multi labels or not.

GoogleCloudDatalabelingV1beta1AnnotationValue
Annotation value for an example.

GoogleCloudDatalabelingV1beta1Attempt
Records a failed evaluation job run.

GoogleCloudDatalabelingV1beta1BigQuerySource
The BigQuery location for input data. If used in an EvaluationJob, this is where the service saves the prediction input and output sampled from the model version.

GoogleCloudDatalabelingV1beta1BoundingBoxEvaluationOptions
Options regarding evaluation between bounding boxes.

GoogleCloudDatalabelingV1beta1BoundingPoly
A bounding polygon in the image.

GoogleCloudDatalabelingV1beta1BoundingPolyConfig
Config for image bounding poly (and bounding box) human labeling task.

GoogleCloudDatalabelingV1beta1ClassificationMetadata
Metadata for classification annotations.

GoogleCloudDatalabelingV1beta1ClassificationMetrics
Metrics calculated for a classification model.

GoogleCloudDatalabelingV1beta1ConfidenceMetricsEntry
There is no detailed description.

GoogleCloudDatalabelingV1beta1ConfusionMatrix
Confusion matrix of the model running the classification. Only applicable when the metrics entry aggregates multiple labels. Not applicable when the entry is for a single label.

GoogleCloudDatalabelingV1beta1ConfusionMatrixEntry
There is no detailed description.

GoogleCloudDatalabelingV1beta1CreateAnnotationSpecSetRequest
Request message for CreateAnnotationSpecSet.

GoogleCloudDatalabelingV1beta1CreateDatasetRequest
Request message for CreateDataset.

GoogleCloudDatalabelingV1beta1CreateEvaluationJobRequest
Request message for CreateEvaluationJob.

GoogleCloudDatalabelingV1beta1CreateInstructionRequest
Request message for CreateInstruction.

GoogleCloudDatalabelingV1beta1CsvInstruction
Deprecated: this instruction format is not supported any more. Instruction from a CSV file.

GoogleCloudDatalabelingV1beta1DataItem
DataItem is a piece of data, without annotation. For example, an image.

GoogleCloudDatalabelingV1beta1Dataset
Dataset is the resource to hold your data. You can request multiple labeling tasks for a dataset while each one will generate an AnnotatedDataset.

GoogleCloudDatalabelingV1beta1Evaluation
Describes an evaluation between a machine learning model's predictions and ground truth labels. Created when an EvaluationJob runs successfully.

GoogleCloudDatalabelingV1beta1EvaluationConfig
Configuration details used for calculating evaluation metrics and creating an Evaluation.

GoogleCloudDatalabelingV1beta1EvaluationJob
Defines an evaluation job that runs periodically to generate Evaluations. Creating an evaluation job is the starting point for using continuous evaluation.

GoogleCloudDatalabelingV1beta1EvaluationJobAlertConfig
Provides details for how an evaluation job sends email alerts based on the results of a run.

GoogleCloudDatalabelingV1beta1EvaluationJobConfig
Configures specific details of how a continuous evaluation job works. Provide this configuration when you create an EvaluationJob.

GoogleCloudDatalabelingV1beta1EvaluationMetrics
There is no detailed description.

GoogleCloudDatalabelingV1beta1EventConfig
Config for video event human labeling task.

GoogleCloudDatalabelingV1beta1Example
An Example is a piece of data and its annotation. For example, an image with label "house".

GoogleCloudDatalabelingV1beta1ExampleComparison
Example comparisons comparing ground truth output and predictions for a specific input.

GoogleCloudDatalabelingV1beta1ExportDataRequest
Request message for ExportData API.

GoogleCloudDatalabelingV1beta1FeedbackMessage
A feedback message inside a feedback thread.

GoogleCloudDatalabelingV1beta1FeedbackThread
A feedback thread of a certain labeling task on a certain annotated dataset.

GoogleCloudDatalabelingV1beta1FeedbackThreadMetadata
There is no detailed description.

GoogleCloudDatalabelingV1beta1GcsDestination
Export destination of the data.Only gcs path is allowed in output_uri.

GoogleCloudDatalabelingV1beta1GcsFolderDestination
Export folder destination of the data.

GoogleCloudDatalabelingV1beta1GcsSource
Source of the Cloud Storage file to be imported.

GoogleCloudDatalabelingV1beta1HumanAnnotationConfig
Configuration for how human labeling task should be done.

GoogleCloudDatalabelingV1beta1ImageBoundingPolyAnnotation
Image bounding poly annotation. It represents a polygon including bounding box in the image.

GoogleCloudDatalabelingV1beta1ImageClassificationAnnotation
Image classification annotation definition.

GoogleCloudDatalabelingV1beta1ImageClassificationConfig
Config for image classification human labeling task.

GoogleCloudDatalabelingV1beta1ImagePayload
Container of information about an image.

GoogleCloudDatalabelingV1beta1ImagePolylineAnnotation
A polyline for the image annotation.

GoogleCloudDatalabelingV1beta1ImageSegmentationAnnotation
Image segmentation annotation.

GoogleCloudDatalabelingV1beta1ImportDataRequest
Request message for ImportData API.

GoogleCloudDatalabelingV1beta1InputConfig
The configuration of input data, including data type, location, etc.

GoogleCloudDatalabelingV1beta1Instruction
Instruction of how to perform the labeling task for human operators. Currently only PDF instruction is supported.

GoogleCloudDatalabelingV1beta1LabelImageRequest
Request message for starting an image labeling task.

GoogleCloudDatalabelingV1beta1LabelStats
Statistics about annotation specs.

GoogleCloudDatalabelingV1beta1LabelTextRequest
Request message for LabelText.

GoogleCloudDatalabelingV1beta1LabelVideoRequest
Request message for LabelVideo.

GoogleCloudDatalabelingV1beta1ListAnnotatedDatasetsResponse
Results of listing annotated datasets for a dataset.

GoogleCloudDatalabelingV1beta1ListAnnotationSpecSetsResponse
Results of listing annotation spec set under a project.

GoogleCloudDatalabelingV1beta1ListDataItemsResponse
Results of listing data items in a dataset.

GoogleCloudDatalabelingV1beta1ListDatasetsResponse
Results of listing datasets within a project.

GoogleCloudDatalabelingV1beta1ListEvaluationJobsResponse
Results for listing evaluation jobs.

GoogleCloudDatalabelingV1beta1ListExamplesResponse
Results of listing Examples in and annotated dataset.

GoogleCloudDatalabelingV1beta1ListFeedbackMessagesResponse
Results for listing FeedbackMessages.

GoogleCloudDatalabelingV1beta1ListFeedbackThreadsResponse
Results for listing FeedbackThreads.

GoogleCloudDatalabelingV1beta1ListInstructionsResponse
Results of listing instructions under a project.

GoogleCloudDatalabelingV1beta1NormalizedBoundingPoly
Normalized bounding polygon.

GoogleCloudDatalabelingV1beta1NormalizedPolyline
Normalized polyline.

GoogleCloudDatalabelingV1beta1NormalizedVertex
A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

GoogleCloudDatalabelingV1beta1ObjectDetectionConfig
Config for video object detection human labeling task. Object detection will be conducted on the images extracted from the video, and those objects will be labeled with bounding boxes. User need to specify the number of images to be extracted per second as the extraction frame rate.

GoogleCloudDatalabelingV1beta1ObjectDetectionMetrics
Metrics calculated for an image object detection (bounding box) model.

GoogleCloudDatalabelingV1beta1ObjectTrackingConfig
Config for video object tracking human labeling task.

GoogleCloudDatalabelingV1beta1ObjectTrackingFrame
Video frame level annotation for object detection and tracking.

GoogleCloudDatalabelingV1beta1OperatorFeedbackMetadata
Metadata describing the feedback from the operator.

GoogleCloudDatalabelingV1beta1OperatorMetadata
General information useful for labels coming from contributors.

GoogleCloudDatalabelingV1beta1OutputConfig
The configuration of output data.

GoogleCloudDatalabelingV1beta1PauseEvaluationJobRequest
Request message for PauseEvaluationJob.

GoogleCloudDatalabelingV1beta1PdfInstruction
Instruction from a PDF file.

GoogleCloudDatalabelingV1beta1Polyline
A line with multiple line segments.

GoogleCloudDatalabelingV1beta1PolylineConfig
Config for image polyline human labeling task.

GoogleCloudDatalabelingV1beta1PrCurve
There is no detailed description.

GoogleCloudDatalabelingV1beta1RequesterFeedbackMetadata
Metadata describing the feedback from the labeling task requester.

GoogleCloudDatalabelingV1beta1ResumeEvaluationJobRequest
Request message ResumeEvaluationJob.

GoogleCloudDatalabelingV1beta1Row
A row in the confusion matrix. Each entry in this row has the same ground truth label.

GoogleCloudDatalabelingV1beta1SearchEvaluationsResponse
Results of searching evaluations.

GoogleCloudDatalabelingV1beta1SearchExampleComparisonsRequest
Request message of SearchExampleComparisons.

GoogleCloudDatalabelingV1beta1SearchExampleComparisonsResponse
Results of searching example comparisons.

GoogleCloudDatalabelingV1beta1SegmentationConfig
Config for image segmentation

GoogleCloudDatalabelingV1beta1SentimentConfig
Config for setting up sentiments.

GoogleCloudDatalabelingV1beta1SequentialSegment
Start and end position in a sequence (e.g. text segment).

GoogleCloudDatalabelingV1beta1TextClassificationAnnotation
Text classification annotation.

GoogleCloudDatalabelingV1beta1TextClassificationConfig
Config for text classification human labeling task.

GoogleCloudDatalabelingV1beta1TextEntityExtractionAnnotation
Text entity extraction annotation.

GoogleCloudDatalabelingV1beta1TextEntityExtractionConfig
Config for text entity extraction human labeling task.

GoogleCloudDatalabelingV1beta1TextMetadata
Metadata for the text.

GoogleCloudDatalabelingV1beta1TextPayload
Container of information about a piece of text.

GoogleCloudDatalabelingV1beta1TimeSegment
A time period inside of an example that has a time dimension (e.g. video).

GoogleCloudDatalabelingV1beta1Vertex
A vertex represents a 2D point in the image. NOTE: the vertex coordinates are in the same scale as the original image.

GoogleCloudDatalabelingV1beta1VideoClassificationAnnotation
Video classification annotation.

GoogleCloudDatalabelingV1beta1VideoClassificationConfig
Config for video classification human labeling task. Currently two types of video classification are supported: 1. Assign labels on the entire video. 2. Split the video into multiple video clips based on camera shot, and assign labels on each video clip.

GoogleCloudDatalabelingV1beta1VideoEventAnnotation
Video event annotation.

GoogleCloudDatalabelingV1beta1VideoObjectTrackingAnnotation
Video object tracking annotation.

GoogleCloudDatalabelingV1beta1VideoPayload
Container of information of a video.

GoogleCloudDatalabelingV1beta1VideoThumbnail
Container of information of a video thumbnail.

GoogleLongrunningListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

ProjectAnnotationSpecSetCreateCall
Creates an annotation spec set by providing a set of labels.

ProjectAnnotationSpecSetDeleteCall
Deletes an annotation spec set by resource name.

ProjectAnnotationSpecSetGetCall
Gets an annotation spec set by resource name.

ProjectAnnotationSpecSetListCall
Lists annotation spec sets for a project. Pagination is supported.

ProjectDatasetAnnotatedDatasetDataItemGetCall
Gets a data item in a dataset by resource name. This API can be called after data are imported into dataset.

ProjectDatasetAnnotatedDatasetDataItemListCall
Lists data items in a dataset. This API can be called after data are imported into dataset. Pagination is supported.

ProjectDatasetAnnotatedDatasetDeleteCall
Deletes an annotated dataset by resource name.

ProjectDatasetAnnotatedDatasetExampleGetCall
Gets an example by resource name, including both data and annotation.

ProjectDatasetAnnotatedDatasetExampleListCall
Lists examples in an annotated dataset. Pagination is supported.

ProjectDatasetAnnotatedDatasetFeedbackThreadDeleteCall
Delete a FeedbackThread.

ProjectDatasetAnnotatedDatasetFeedbackThreadFeedbackMessageCreateCall
Create a FeedbackMessage object.

ProjectDatasetAnnotatedDatasetFeedbackThreadFeedbackMessageDeleteCall
Delete a FeedbackMessage.

ProjectDatasetAnnotatedDatasetFeedbackThreadFeedbackMessageGetCall
Get a FeedbackMessage object.

ProjectDatasetAnnotatedDatasetFeedbackThreadFeedbackMessageListCall
List FeedbackMessages with pagination.

ProjectDatasetAnnotatedDatasetFeedbackThreadGetCall
Get a FeedbackThread object.

ProjectDatasetAnnotatedDatasetFeedbackThreadListCall
List FeedbackThreads with pagination.

ProjectDatasetAnnotatedDatasetGetCall
Gets an annotated dataset by resource name.

ProjectDatasetAnnotatedDatasetListCall
Lists annotated datasets for a dataset. Pagination is supported.

ProjectDatasetCreateCall
Creates dataset. If success return a Dataset resource.

ProjectDatasetDataItemGetCall
Gets a data item in a dataset by resource name. This API can be called after data are imported into dataset.

ProjectDatasetDataItemListCall
Lists data items in a dataset. This API can be called after data are imported into dataset. Pagination is supported.

ProjectDatasetDeleteCall
Deletes a dataset by resource name.

ProjectDatasetEvaluationExampleComparisonSearchCall
Searches example comparisons from an evaluation. The return format is a list of example comparisons that show ground truth and prediction(s) for a single input. Search by providing an evaluation ID.

ProjectDatasetEvaluationGetCall
Gets an evaluation by resource name (to search, use projects.evaluations.search).

ProjectDatasetExportDataCall
Exports data and annotations from dataset.

ProjectDatasetGetCall
Gets dataset by resource name.

ProjectDatasetImageLabelCall
Starts a labeling task for image. The type of image labeling task is configured by feature in the request.

ProjectDatasetImportDataCall
Imports data into dataset based on source locations defined in request. It can be called multiple times for the same dataset. Each dataset can only have one long running operation running on it. For example, no labeling task (also long running operation) can be started while importing is still ongoing. Vice versa.

ProjectDatasetListCall
Lists datasets under a project. Pagination is supported.

ProjectDatasetTextLabelCall
Starts a labeling task for text. The type of text labeling task is configured by feature in the request.

ProjectDatasetVideoLabelCall
Starts a labeling task for video. The type of video labeling task is configured by feature in the request.

ProjectEvaluationJobCreateCall
Creates an evaluation job.

ProjectEvaluationJobDeleteCall
Stops and deletes an evaluation job.

ProjectEvaluationJobGetCall
Gets an evaluation job by resource name.

ProjectEvaluationJobListCall
Lists all evaluation jobs within a project with possible filters. Pagination is supported.

ProjectEvaluationJobPatchCall
Updates an evaluation job. You can only update certain fields of the job's EvaluationJobConfig: humanAnnotationConfig.instruction, exampleCount, and exampleSamplePercentage. If you want to change any other aspect of the evaluation job, you must delete the job and create a new one.

ProjectEvaluationJobPauseCall
Pauses an evaluation job. Pausing an evaluation job that is already in a PAUSED state is a no-op.

ProjectEvaluationJobResumeCall
Resumes a paused evaluation job. A deleted evaluation job can't be resumed. Resuming a running or scheduled evaluation job is a no-op.

ProjectEvaluationSearchCall
Searches evaluations within a project.

ProjectInstructionCreateCall
Creates an instruction for how data should be labeled.

ProjectInstructionDeleteCall
Deletes an instruction object by resource name.

ProjectInstructionGetCall
Gets an instruction by resource name.

ProjectInstructionListCall
Lists instructions for a project. Pagination is supported.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the DataLabeling hub.

ProjectOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.