Struct google_datalabeling1_beta1::api::GoogleCloudDatalabelingV1beta1EvaluationJob

pub struct GoogleCloudDatalabelingV1beta1EvaluationJob {
    pub annotation_spec_set: Option<String>,
    pub attempts: Option<Vec<GoogleCloudDatalabelingV1beta1Attempt>>,
    pub create_time: Option<String>,
    pub description: Option<String>,
    pub evaluation_job_config: Option<GoogleCloudDatalabelingV1beta1EvaluationJobConfig>,
    pub label_missing_ground_truth: Option<bool>,
    pub model_version: Option<String>,
    pub name: Option<String>,
    pub schedule: Option<String>,
    pub state: Option<String>,
}

Defines an evaluation job that runs periodically to generate Evaluations. Creating an evaluation job is the starting point for using continuous evaluation.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

evaluation jobs create projects (response)

evaluation jobs get projects (response)

evaluation jobs patch projects (request|response)

Fields
annotation_spec_set: Option<String>

Required. Name of the AnnotationSpecSet describing all the labels that your machine learning model outputs. You must create this resource before you create an evaluation job and provide its name in the following format: "projects/{project_id}/annotationSpecSets/{annotation_spec_set_id}"

attempts: Option<Vec<GoogleCloudDatalabelingV1beta1Attempt>>

Output only. Every time the evaluation job runs and an error occurs, the failed attempt is appended to this array.

create_time: Option<String>

Output only. Timestamp of when this evaluation job was created.

description: Option<String>

Required. Description of the job. The description can be up to 25,000 characters long.

evaluation_job_config: Option<GoogleCloudDatalabelingV1beta1EvaluationJobConfig>

Required. Configuration details for the evaluation job.

label_missing_ground_truth: Option<bool>

Required. Whether you want Data Labeling Service to provide ground truth labels for prediction input. If you want the service to assign human labelers to annotate your data, set this to true. If you want to provide your own ground truth labels in the evaluation job's BigQuery table, set this to false.

model_version: Option<String>

Required. The AI Platform Prediction model version to be evaluated. Prediction input and output is sampled from this model version. When creating an evaluation job, specify the model version in the following format: "projects/{project_id}/models/{model_name}/versions/{version_name}" There can only be one evaluation job per model version.

name: Option<String>

Output only. After you create a job, Data Labeling Service assigns a name to the job with the following format: "projects/{project_id}/evaluationJobs/ {evaluation_job_id}"

schedule: Option<String>

Required. Describes the interval at which the job runs. This interval must be at least 1 day, and it is rounded to the nearest day. For example, if you specify a 50-hour interval, the job runs every 2 days. You can provide the schedule in crontab format or in an English-like format. Regardless of what you specify, the job will run at 10:00 AM UTC. Only the interval from this schedule is used, not the specific time of day.

state: Option<String>

Output only. Describes the current state of the job.