Struct google_datalabeling1_beta1::api::GoogleCloudDatalabelingV1beta1EvaluationJobConfig

pub struct GoogleCloudDatalabelingV1beta1EvaluationJobConfig {
    pub bigquery_import_keys: Option<HashMap<String, String>>,
    pub bounding_poly_config: Option<GoogleCloudDatalabelingV1beta1BoundingPolyConfig>,
    pub evaluation_config: Option<GoogleCloudDatalabelingV1beta1EvaluationConfig>,
    pub evaluation_job_alert_config: Option<GoogleCloudDatalabelingV1beta1EvaluationJobAlertConfig>,
    pub example_count: Option<i32>,
    pub example_sample_percentage: Option<f64>,
    pub human_annotation_config: Option<GoogleCloudDatalabelingV1beta1HumanAnnotationConfig>,
    pub image_classification_config: Option<GoogleCloudDatalabelingV1beta1ImageClassificationConfig>,
    pub input_config: Option<GoogleCloudDatalabelingV1beta1InputConfig>,
    pub text_classification_config: Option<GoogleCloudDatalabelingV1beta1TextClassificationConfig>,
}

Configures specific details of how a continuous evaluation job works. Provide this configuration when you create an EvaluationJob.

This type is not used in any activity, and only used as part of another schema.

Fields
bigquery_import_keys: Option<HashMap<String, String>>

Required. Prediction keys that tell Data Labeling Service where to find the data for evaluation in your BigQuery table. When the service samples prediction input and output from your model version and saves it to BigQuery, the data gets stored as JSON strings in the BigQuery table. These keys tell Data Labeling Service how to parse the JSON. You can provide the following entries in this field: * data_json_key: the data key for prediction input. You must provide either this key or reference_json_key. * reference_json_key: the data reference key for prediction input. You must provide either this key or data_json_key. * label_json_key: the label key for prediction output. Required. * label_score_json_key: the score key for prediction output. Required. * bounding_box_json_key: the bounding box key for prediction output. Required if your model version perform image object detection. Learn how to configure prediction keys.

bounding_poly_config: Option<GoogleCloudDatalabelingV1beta1BoundingPolyConfig>

Specify this field if your model version performs image object detection (bounding box detection). annotationSpecSet in this configuration must match EvaluationJob.annotationSpecSet.

evaluation_config: Option<GoogleCloudDatalabelingV1beta1EvaluationConfig>

Required. Details for calculating evaluation metrics and creating Evaulations. If your model version performs image object detection, you must specify the boundingBoxEvaluationOptions field within this configuration. Otherwise, provide an empty object for this configuration.

evaluation_job_alert_config: Option<GoogleCloudDatalabelingV1beta1EvaluationJobAlertConfig>

Optional. Configuration details for evaluation job alerts. Specify this field if you want to receive email alerts if the evaluation job finds that your predictions have low mean average precision during a run.

example_count: Option<i32>

Required. The maximum number of predictions to sample and save to BigQuery during each evaluation interval. This limit overrides example_sample_percentage: even if the service has not sampled enough predictions to fulfill example_sample_perecentage during an interval, it stops sampling predictions when it meets this limit.

example_sample_percentage: Option<f64>

Required. Fraction of predictions to sample and save to BigQuery during each evaluation interval. For example, 0.1 means 10% of predictions served by your model version get saved to BigQuery.

human_annotation_config: Option<GoogleCloudDatalabelingV1beta1HumanAnnotationConfig>

Optional. Details for human annotation of your data. If you set labelMissingGroundTruth to true for this evaluation job, then you must specify this field. If you plan to provide your own ground truth labels, then omit this field. Note that you must create an Instruction resource before you can specify this field. Provide the name of the instruction resource in the instruction field within this configuration.

image_classification_config: Option<GoogleCloudDatalabelingV1beta1ImageClassificationConfig>

Specify this field if your model version performs image classification or general classification. annotationSpecSet in this configuration must match EvaluationJob.annotationSpecSet. allowMultiLabel in this configuration must match classificationMetadata.isMultiLabel in input_config.

input_config: Option<GoogleCloudDatalabelingV1beta1InputConfig>

Rquired. Details for the sampled prediction input. Within this configuration, there are requirements for several fields: * dataType must be one of IMAGE, TEXT, or GENERAL_DATA. * annotationType must be one of IMAGE_CLASSIFICATION_ANNOTATION, TEXT_CLASSIFICATION_ANNOTATION, GENERAL_CLASSIFICATION_ANNOTATION, or IMAGE_BOUNDING_BOX_ANNOTATION (image object detection). * If your machine learning model performs classification, you must specify classificationMetadata.isMultiLabel. * You must specify bigquerySource (not gcsSource).

text_classification_config: Option<GoogleCloudDatalabelingV1beta1TextClassificationConfig>

Specify this field if your model version performs text classification. annotationSpecSet in this configuration must match EvaluationJob.annotationSpecSet. allowMultiLabel in this configuration must match classificationMetadata.isMultiLabel in input_config.